requirements.txt
# CrowdWisdomTrading SEC Chat AI Agent - Requirements

# LLM Provider (OpenRouter uses OpenAI-compatible API)
openai>=1.0.0

# HTTP & Async
httpx>=0.27.0
aiohttp>=3.9.0
requests>=2.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# SEC Data (use edgar, NOT sec-api which needs a paid key)
edgar>=5.4.0

# Apify Client (Twitter/X scraping)
apify-client>=1.6.0

# RAG / Vector Store
chromadb>=0.5.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Charting
matplotlib>=3.7.0
plotly>=5.17.0

# Environment
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
rich>=13.0.0
tenacity>=8.2.0